+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [1.9.0] - 2026-05-31
+
 ### Added
 
 - Add support for C++ module files.
@@ -154,7 +156,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 First stable release.
 
-[unreleased]: https://github.com/chausner/vcxproj2cmake/compare/v1.8.0...HEAD
+[unreleased]: https://github.com/chausner/vcxproj2cmake/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.5.0...v1.6.0